Restructure client
This is a major restructuring of the client code to a new architecture.

Instead of having two clients that implement the `Client` trait we only
have one `Client` with two backends—one for the remote node the other
for the emulator (previously `MemoryClient`).

Follow-up
* Turn `ClientWithExecutor` into a backend, if possible
* Eliminate `Client` trait and only provide `Client` structure to
  consumers.
